refactor(get-queries-for-element): convert to TS
